+++ ext/fs/lib.rs
@@ -95,7 +95,6 @@ deno_core::extension!(deno_fs,
     op_fs_remove_async<P>,
     op_fs_copy_file_sync<P>,
     op_fs_copy_file_async<P>,
-    op_fs_exists_sync<P>,
     op_fs_stat_sync<P>,
     op_fs_stat_async<P>,
     op_fs_lstat_sync<P>,
