+++ ext/node/ops/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
 pub mod crypto;
+pub mod fs;
 pub mod http;
 pub mod http2;
 pub mod idna;
